+++ src/Soluble/FlexStore/Exception/EmptyQueryException.php
@@ -1,7 +1,13 @@
 <?php
 
-/**
- * @author Vanvelthem Sébastien
+/*
+ * soluble-flexstore library
+ *
+ * @author    Vanvelthem Sébastien
+ * @link      https://github.com/belgattitude/soluble-flexstore
+ * @copyright Copyright (c) 20016-2017 Vanvelthem Sébastien
+ * @license   MIT License https://github.com/belgattitude/soluble-flexstore/blob/master/LICENSE.md
+ *
  */
 
 namespace Soluble\FlexStore\Exception;