+++ src/Soluble/FlexStore/Formatter/CurrencyFormatter.php
@@ -1,5 +1,15 @@
 <?php
 
+/*
+ * soluble-flexstore library
+ *
+ * @author    Vanvelthem Sébastien
+ * @link      https://github.com/belgattitude/soluble-flexstore
+ * @copyright Copyright (c) 20016-2017 Vanvelthem Sébastien
+ * @license   MIT License https://github.com/belgattitude/soluble-flexstore/blob/master/LICENSE.md
+ *
+ */
+
 namespace Soluble\FlexStore\Formatter;
 
 use Soluble\FlexStore\Exception;
@@ -122,7 +132,7 @@ class CurrencyFormatter extends NumberFormatter
         $result = $this->formatters[$formatterId]->parseCurrency($value, $currency);
 
         if ($value === false) {
-            return;
+            return null;
         }
 
         return ['value' => $result, 'currency' => $currency];