+++ src/Soluble/FlexStore/Source/AbstractSource.php
@@ -1,6 +1,13 @@
 <?php
-/**
- * @author Vanvelthem Sébastien
+
+/*
+ * soluble-flexstore library
+ *
+ * @author    Vanvelthem Sébastien
+ * @link      https://github.com/belgattitude/soluble-flexstore
+ * @copyright Copyright (c) 20016-2017 Vanvelthem Sébastien
+ * @license   MIT License https://github.com/belgattitude/soluble-flexstore/blob/master/LICENSE.md
+ *
  */
 
 namespace Soluble\FlexStore\Source;
@@ -24,7 +31,7 @@ abstract class AbstractSource implements SourceInterface
     protected $columnModel;
 
     /**
-     * @var string|int
+     * @var string|null
      */
     protected $identifier;
 
@@ -55,7 +62,7 @@ abstract class AbstractSource implements SourceInterface
     /**
      * Set the primary key / unique identifier in the store.
      *
-     * @param string|null $identifier column name of the primary key
+     * @param string $identifier column name of the primary key
      *
      * @return AbstractSource
      */