+++ src/Soluble/FlexStore/Source/QueryableSourceInterface.php
@@ -1,13 +1,17 @@
 <?php
 
+/*
+ * soluble-flexstore library
+ *
+ * @author    Vanvelthem Sébastien
+ * @link      https://github.com/belgattitude/soluble-flexstore
+ * @copyright Copyright (c) 20016-2017 Vanvelthem Sébastien
+ * @license   MIT License https://github.com/belgattitude/soluble-flexstore/blob/master/LICENSE.md
+ *
+ */
+
 namespace Soluble\FlexStore\Source;
 
-interface QueryableSourceInterface
+interface QueryableSourceInterface extends SourceInterface
 {
-    /**
-     * Return underlying query (sql) string.
-     *
-     * @return string
-     */
-    public function getQueryString();
 }