+++ src/Soluble/FlexStore/Store.php
@@ -1,9 +1,13 @@
 <?php
 
-/**
- * @author Vanvelthem Sébastien
+/*
+ * soluble-flexstore library
  *
- * @deprecated
+ * @author    Vanvelthem Sébastien
+ * @link      https://github.com/belgattitude/soluble-flexstore
+ * @copyright Copyright (c) 20016-2017 Vanvelthem Sébastien
+ * @license   MIT License https://github.com/belgattitude/soluble-flexstore/blob/master/LICENSE.md
+ *
  */
 
 namespace Soluble\FlexStore;