+++ src/Soluble/FlexStore/Store/StoreInterface.php
@@ -1,7 +1,13 @@
 <?php
 
-/**
- * @author Vanvelthem Sébastien
+/*
+ * soluble-flexstore library
+ *
+ * @author    Vanvelthem Sébastien
+ * @link      https://github.com/belgattitude/soluble-flexstore
+ * @copyright Copyright (c) 20016-2017 Vanvelthem Sébastien
+ * @license   MIT License https://github.com/belgattitude/soluble-flexstore/blob/master/LICENSE.md
+ *
  */
 
 namespace Soluble\FlexStore\Store;
@@ -30,4 +36,9 @@ interface StoreInterface
      * @return \Soluble\FlexStore\Column\ColumnModel
      */
     public function getColumnModel();
+
+    /**
+     * @return \Soluble\FlexStore\Options
+     */
+    public function getOptions();
 }