+++ src/Soluble/FlexStore/Writer/Excel/Exception/ExtensionNotLoadedException.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * soluble-flexstore library
+ *
+ * @author    Vanvelthem Sébastien
+ * @link      https://github.com/belgattitude/soluble-flexstore
+ * @copyright Copyright (c) 20016-2017 Vanvelthem Sébastien
+ * @license   MIT License https://github.com/belgattitude/soluble-flexstore/blob/master/LICENSE.md
+ *
+ */
+
+namespace Soluble\FlexStore\Writer\Excel\Exception;
+
+use DomainException;
+
+class ExtensionNotLoadedException extends DomainException
+{
+}