+++ src/Soluble/FlexStore/Writer/Zend/JsonModelWriter.php
@@ -1,5 +1,15 @@
 <?php
 
+/*
+ * soluble-flexstore library
+ *
+ * @author    Vanvelthem Sébastien
+ * @link      https://github.com/belgattitude/soluble-flexstore
+ * @copyright Copyright (c) 20016-2017 Vanvelthem Sébastien
+ * @license   MIT License https://github.com/belgattitude/soluble-flexstore/blob/master/LICENSE.md
+ *
+ */
+
 namespace Soluble\FlexStore\Writer\Zend;
 
 use Zend\View\Model\JsonModel as ZendJsonModel;
@@ -12,7 +22,7 @@ class JsonModelWriter extends AbstractWriter
     /**
      * @param Options $options
      *
-     * @return \Zend\View\Model\JsonModel
+     * @return ZendJsonModel
      */
     public function getData(Options $options = null)
     {