+++ src/Soluble/FlexStore/Writer/Zend/JsonWriter.php
@@ -1,5 +1,15 @@
 <?php
 
+/*
+ * soluble-flexstore library
+ *
+ * @author    Vanvelthem Sébastien
+ * @link      https://github.com/belgattitude/soluble-flexstore
+ * @copyright Copyright (c) 20016-2017 Vanvelthem Sébastien
+ * @license   MIT License https://github.com/belgattitude/soluble-flexstore/blob/master/LICENSE.md
+ *
+ */
+
 namespace Soluble\FlexStore\Writer\Zend;
 
 use Soluble\FlexStore\Writer\Http\SimpleHeaders;
@@ -18,7 +28,7 @@ class JsonWriter extends AbstractSendableWriter
     /**
      * @param Options $options
      *
-     * @return \Zend\View\Model\JsonModel
+     * @return string
      */
     public function getData(Options $options = null)
     {