+++ tests/bootstrap.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * soluble-flexstore library
+ *
+ * @author    Vanvelthem Sébastien
+ * @link      https://github.com/belgattitude/soluble-flexstore
+ * @copyright Copyright (c) 20016-2017 Vanvelthem Sébastien
+ * @license   MIT License https://github.com/belgattitude/soluble-flexstore/blob/master/LICENSE.md
+ *
+ */
+
+if (!$loader = @include __DIR__ . '/../vendor/autoload.php') {
+    die('You must set up the project dependencies, run the following commands:' . PHP_EOL .
+            'curl -s http://getcomposer.org/installer | php' . PHP_EOL .
+            'php composer.phar install' . PHP_EOL);
+}
+
+ini_set('error_reporting', E_ALL | E_STRICT);
+$baseDir = dirname(__DIR__);
+
+require __DIR__ . '/../vendor/autoload.php';